+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# 0.0.0 (2021-06-15)
+# 0.0.0 (2021-06-16)
 
 ### Bug Fixes
 
@@ -41,10 +41,11 @@
 - **algorithm:** category ([bc1726f](https://github.com/air-supply94/js-algorithm/commit/bc1726fe8bad0c41b79600ade1c12c98052e1b20))
 - **algorithm/array:** add randomizedSet ([8ffb01a](https://github.com/air-supply94/js-algorithm/commit/8ffb01a022d03f60187cf3b88184867a9ad26249))
 - **algorithm/array:** add randomizedSetBlackList ([56ab592](https://github.com/air-supply94/js-algorithm/commit/56ab59276fbda3d24b0612b5abc68ab7bb055810))
-- **algorithm/array:** add removeDuplicateLetters ([94ba9df](https://github.com/air-supply94/js-algorithm/commit/94ba9dfd2359d6ab987723b12170be91db1a610d))
+- **algorithm/array:** add removeDuplicateLetters ([e258403](https://github.com/air-supply94/js-algorithm/commit/e25840313d53394781e69f1abd36c01d95ee35ae))
 - **algorithm/array:** add removeDuplicates ([e3412f8](https://github.com/air-supply94/js-algorithm/commit/e3412f86da8aa337a17215ea222fa772978b91e6))
 - **algorithm/array:** add removeElement ([6940656](https://github.com/air-supply94/js-algorithm/commit/69406564fc87b8918ecef93a4e2feea2677e6489))
 - **algorithm/array:** add twoSum ([a93da8f](https://github.com/air-supply94/js-algorithm/commit/a93da8fdd1a5cf9228a8cda70410b1e47433fc85))
+- **algorithm/backTrack:** add combine ([6f681c8](https://github.com/air-supply94/js-algorithm/commit/6f681c83ce8ef381ac6efccd8481673453cfcd5d))
 - **algorithm/classic:** add coinChange ([4c2b923](https://github.com/air-supply94/js-algorithm/commit/4c2b92300783066c27c01ce3f79989fbdfcf0314))
 - **algorithm/classic:** add maximalSquare ([82da714](https://github.com/air-supply94/js-algorithm/commit/82da714f99bc9b61b8b2498c1bac7e0900b08540))
 - **algorithm/classic:** add maxSubArray ([684c1b8](https://github.com/air-supply94/js-algorithm/commit/684c1b823a8541d96bbbb4e245ef456e09425e35))